2003-01-15  Pierre-Luc Paour <gallery@paour.com> (1.3-b16)
	* MASSIVE refactoring to make GR more modular to prepare
	  for the various applet versions.
	* Changed log location from working directory to temp dir.
	* Added a mini applet that only contains a list of files
	  and buttons to add pictures and upload.

+++ com/gallery/GalleryRemote/prefs/PropertiesFile.java
@@ -37,6 +37,7 @@ public class PropertiesFile extends GalleryProperties {
 	protected boolean read = false;
 	protected boolean written = false;
 	protected String mFilename;
+	protected boolean readOnly = false;
 
 	/**
 	 * Constructor for the PropertiesFile object
@@ -97,6 +98,10 @@ public class PropertiesFile extends GalleryProperties {
 		} else {
 			mFilename = name;
 		}
+	}
+
+	public void setReadOnly() {
+		readOnly = true;
 	}
 
 
@@ -181,7 +186,12 @@ public class PropertiesFile extends GalleryProperties {
 	 * Write the property file to disk
 	 */
 	public synchronized void write() {
-		if (!written) {
+		if (readOnly && defaults != null && defaults instanceof PropertiesFile) {
+			((PropertiesFile) defaults).write();
+			written = true;
+		}
+
+		if (!written && !readOnly) {
 			FileOutputStream fileOut = null;
 			try {
 				fileOut = new FileOutputStream(mFilename);
